Make grpc server reusable by other metadata store implementations
