+++ crates/metadata-store/src/grpc/client.rs
@@ -21,17 +21,17 @@ use restate_core::network::net_util::CommonClientConnectionOptions;
 use restate_types::net::AdvertisedAddress;
 use restate_types::Version;
 
+use crate::grpc::pb_conversions::ConversionError;
 use crate::grpc_svc::metadata_store_svc_client::MetadataStoreSvcClient;
 use crate::grpc_svc::{DeleteRequest, GetRequest, PutRequest};
-use crate::local::grpc::pb_conversions::ConversionError;
 
-/// Client end to interact with the [`LocalMetadataStore`].
+/// Client end to interact with the metadata store.
 #[derive(Debug, Clone)]
-pub struct LocalMetadataStoreClient {
+pub struct GrpcMetadataStoreClient {
     svc_client: MetadataStoreSvcClient<Channel>,
 }
 
-impl LocalMetadataStoreClient {
+impl GrpcMetadataStoreClient {
     pub fn new<T: CommonClientConnectionOptions>(
         metadata_store_address: AdvertisedAddress,
         options: &T,
@@ -45,7 +45,7 @@ impl LocalMetadataStoreClient {
 }
 
 #[async_trait]
-impl ProvisionedMetadataStore for LocalMetadataStoreClient {
+impl ProvisionedMetadataStore for GrpcMetadataStoreClient {
     async fn get(&self, key: ByteString) -> Result<Option<VersionedValue>, ReadError> {
         let response = self
             .svc_client
