+++ crates/metadata-store/src/lib.rs
@@ -8,9 +8,70 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+mod grpc;
 mod grpc_svc;
 pub mod local;
 
+use bytestring::ByteString;
+use restate_core::metadata_store::VersionedValue;
 pub use restate_core::metadata_store::{
     MetadataStoreClient, Precondition, ReadError, ReadModifyWriteError, WriteError,
 };
+use restate_types::errors::GenericError;
+use restate_types::storage::{StorageDecodeError, StorageEncodeError};
+use restate_types::Version;
+use tokio::sync::{mpsc, oneshot};
+
+pub type RequestSender = mpsc::Sender<MetadataStoreRequest>;
+pub type RequestReceiver = mpsc::Receiver<MetadataStoreRequest>;
+
+type Result<T, E = Error> = std::result::Result<T, E>;
+
+#[derive(Debug)]
+pub enum MetadataStoreRequest {
+    Get {
+        key: ByteString,
+        result_tx: oneshot::Sender<Result<Option<VersionedValue>>>,
+    },
+    GetVersion {
+        key: ByteString,
+        result_tx: oneshot::Sender<Result<Option<Version>>>,
+    },
+    Put {
+        key: ByteString,
+        value: VersionedValue,
+        precondition: Precondition,
+        result_tx: oneshot::Sender<Result<()>>,
+    },
+    Delete {
+        key: ByteString,
+        precondition: Precondition,
+        result_tx: oneshot::Sender<Result<()>>,
+    },
+}
+
+#[derive(Debug, thiserror::Error)]
+pub enum Error {
+    #[error("storage error: {0}")]
+    Storage(#[from] GenericError),
+    #[error("failed precondition: {0}")]
+    FailedPrecondition(String),
+    #[error("invalid argument: {0}")]
+    InvalidArgument(String),
+    #[error("encode error: {0}")]
+    Encode(#[from] StorageEncodeError),
+    #[error("decode error: {0}")]
+    Decode(#[from] StorageDecodeError),
+}
+
+impl Error {
+    fn kv_pair_exists() -> Self {
+        Error::FailedPrecondition("key-value pair already exists".to_owned())
+    }
+
+    fn version_mismatch(expected: Version, actual: Option<Version>) -> Self {
+        Error::FailedPrecondition(format!(
+            "Expected version '{expected}' but found version '{actual:?}'"
+        ))
+    }
+}
